refactor: rename server db migrations in preparation for client migrations
