+++ misc/git-hooks/pre-commit
@@ -147,7 +147,10 @@ function check_trailing_whitespace() {
       >&2 echo "Warning: no commits yet, checking against --root"
       rev="--root"
     fi
-    if ! git diff --check $rev ; then
+    # Exclude db migration files
+    if git diff --check $rev | \
+        grep -v -E '^diff --git a/.*b/db/' |
+        grep -E '^-.*db/' ; then
       >&2 echo "Trailing whitespace detected. Please remove them before committing."
       return 1
     fi
