+++ modules/fedimint-ln-server/Cargo.toml
@@ -49,4 +49,3 @@ assert_matches = "1.5.0"
 test-log = { version = "0.2", features = [ "trace" ], default-features = false }
 tokio = {version = "1.26.0", features = [ "full" ] }
 tracing-subscriber = { version = "0.3.16", features = [ "env-filter" ] }
-fedimint-testing = { path = "../../fedimint-testing" }
