+++ modules/fedimint-ln-tests/Cargo.toml
@@ -15,6 +15,7 @@ path = "tests/tests.rs"
 anyhow = "1.0.66"
 assert_matches = "1.5.0"
 bitcoin = { version = "0.29.2", features = [ "rand", "serde"] }
+bitcoin_hashes = "0.11.0"
 fedimint-bitcoind = { path = "../../fedimint-bitcoind" }
 fedimint-dummy-common = { path = "../fedimint-dummy-common" }
 fedimint-dummy-client = { path = "../fedimint-dummy-client" }
@@ -27,7 +28,13 @@ fedimint-client = { path = "../../fedimint-client" }
 fedimint-core ={ path = "../../fedimint-core" }
 fedimint-server = { path = "../../fedimint-server" }
 fedimint-logging = { path = "../../fedimint-logging" }
+futures = "0.3.24"
 lightning-invoice = { version = "0.26.0", features = [ "serde" ] }
+threshold_crypto = { workspace = true }
 tokio = { version = "1.26.0", features = ["sync"] }
 tracing = "0.1.37"
+secp256k1 = { version="0.24.2", default-features=false }
 serde_json = "1.0.91"
+rand = "0.8"
+strum = "0.24"
+strum_macros = "0.24"
