+++ modules/fedimint-mint-tests/Cargo.toml
@@ -13,6 +13,7 @@ path = "tests/tests.rs"
 
 [dependencies]
 anyhow = "1.0.66"
+bitcoin_hashes = "0.11.0"
 fedimint-dummy-common = { path = "../fedimint-dummy-common" }
 fedimint-dummy-client = { path = "../fedimint-dummy-client" }
 fedimint-dummy-server = { path = "../fedimint-dummy-server" }
@@ -24,5 +25,11 @@ fedimint-client = { path = "../../fedimint-client" }
 fedimint-core ={ path = "../../fedimint-core" }
 fedimint-server = { path = "../../fedimint-server" }
 fedimint-logging = { path = "../../fedimint-logging" }
+futures = "0.3"
 tokio = { version = "1.26.0", features = ["sync"] }
 tracing = "0.1.37"
+rand = "0.8"
+secp256k1 = "0.24.2"
+tbs = { package = "fedimint-tbs", version = "0.3.0-alpha", path = "../../crypto/tbs" }
+strum = "0.24"
+strum_macros = "0.24"
