+++ modules/fedimint-wallet-server/src/lib.rs
@@ -1706,348 +1706,3 @@ mod tests {
         })
     }
 }
-
-#[cfg(test)]
-mod fedimint_migration_tests {
-    use anyhow::{ensure, Context};
-    use bitcoin::psbt::{Input, PartiallySignedTransaction};
-    use bitcoin::{
-        Amount, BlockHash, PackedLockTime, Script, Sequence, Transaction, TxIn, TxOut, Txid,
-        WPubkeyHash,
-    };
-    use fedimint_core::core::LEGACY_HARDCODED_INSTANCE_ID_WALLET;
-    use fedimint_core::db::{
-        apply_migrations, DatabaseTransaction, IDatabaseTransactionOpsCoreTyped,
-    };
-    use fedimint_core::module::registry::ModuleDecoderRegistry;
-    use fedimint_core::module::{CommonModuleInit, DynServerModuleInit};
-    use fedimint_core::{BitcoinHash, Feerate, OutPoint, PeerId, ServerModule, TransactionId};
-    use fedimint_testing::db::{
-        prepare_db_migration_snapshot, validate_migrations, BYTE_20, BYTE_32, BYTE_33,
-    };
-    use fedimint_wallet_common::db::{
-        BlockCountVoteKey, BlockCountVotePrefix, BlockHashKey, BlockHashKeyPrefix, DbKeyPrefix,
-        FeeRateVoteKey, FeeRateVotePrefix, PegOutBitcoinTransaction,
-        PegOutBitcoinTransactionPrefix, PegOutNonceKey, PegOutTxSignatureCI,
-        PegOutTxSignatureCIPrefix, PendingTransactionKey, PendingTransactionPrefixKey, UTXOKey,
-        UTXOPrefixKey, UnsignedTransactionKey, UnsignedTransactionPrefixKey,
-    };
-    use fedimint_wallet_common::{
-        PegOutFees, PendingTransaction, Rbf, SpendableUTXO, UnsignedTransaction, WalletCommonInit,
-        WalletOutputOutcome,
-    };
-    use futures::StreamExt;
-    use rand::rngs::OsRng;
-    use secp256k1::Message;
-    use strum::IntoEnumIterator;
-
-    use crate::{Wallet, WalletInit};
-
-    /// Create a database with version 0 data. The database produced is not
-    /// intended to be real data or semantically correct. It is only
-    /// intended to provide coverage when reading the database
-    /// in future code versions. This function should not be updated when
-    /// database keys/values change - instead a new function should be added
-    /// that creates a new database backup that can be tested.
-    async fn create_db_with_v0_data(mut dbtx: DatabaseTransaction<'_>) {
-        dbtx.insert_new_entry(&BlockHashKey(BlockHash::from_slice(&BYTE_32).unwrap()), &())
-            .await;
-
-        let utxo = UTXOKey(bitcoin::OutPoint {
-            txid: Txid::from_slice(&BYTE_32).unwrap(),
-            vout: 0,
-        });
-        let spendable_utxo = SpendableUTXO {
-            tweak: BYTE_33,
-            amount: Amount::from_sat(10000),
-        };
-
-        dbtx.insert_new_entry(&utxo, &spendable_utxo).await;
-
-        dbtx.insert_new_entry(&PegOutNonceKey, &1).await;
-
-        dbtx.insert_new_entry(&BlockCountVoteKey(PeerId::from(0)), &1)
-            .await;
-
-        dbtx.insert_new_entry(
-            &FeeRateVoteKey(PeerId::from(0)),
-            &Feerate { sats_per_kvb: 10 },
-        )
-        .await;
-
-        let unsigned_transaction_key = UnsignedTransactionKey(Txid::from_slice(&BYTE_32).unwrap());
-
-        let selected_utxos: Vec<(UTXOKey, SpendableUTXO)> = vec![(utxo.clone(), spendable_utxo)];
-
-        let destination = Script::new_v0_p2wpkh(&WPubkeyHash::from_slice(&BYTE_20).unwrap());
-        let output: Vec<TxOut> = vec![TxOut {
-            value: 10000,
-            script_pubkey: destination.clone(),
-        }];
-
-        let transaction = Transaction {
-            version: 2,
-            lock_time: PackedLockTime::ZERO,
-            input: vec![TxIn {
-                previous_output: utxo.0,
-                script_sig: Default::default(),
-                sequence: Sequence::ENABLE_RBF_NO_LOCKTIME,
-                witness: bitcoin::Witness::new(),
-            }],
-            output,
-        };
-
-        let inputs = vec![Input {
-            non_witness_utxo: None,
-            witness_utxo: Some(TxOut {
-                value: 10000,
-                script_pubkey: destination.clone(),
-            }),
-            partial_sigs: Default::default(),
-            sighash_type: None,
-            redeem_script: None,
-            witness_script: Some(destination.clone()),
-            bip32_derivation: Default::default(),
-            final_script_sig: None,
-            final_script_witness: None,
-            ripemd160_preimages: Default::default(),
-            sha256_preimages: Default::default(),
-            hash160_preimages: Default::default(),
-            hash256_preimages: Default::default(),
-            proprietary: Default::default(),
-            tap_key_sig: Default::default(),
-            tap_script_sigs: Default::default(),
-            tap_scripts: Default::default(),
-            tap_key_origins: Default::default(),
-            tap_internal_key: Default::default(),
-            tap_merkle_root: Default::default(),
-            unknown: Default::default(),
-        }];
-
-        let psbt = PartiallySignedTransaction {
-            unsigned_tx: transaction.clone(),
-            version: 0,
-            xpub: Default::default(),
-            proprietary: Default::default(),
-            unknown: Default::default(),
-            inputs,
-            outputs: vec![Default::default()],
-        };
-
-        let unsigned_transaction = UnsignedTransaction {
-            psbt,
-            signatures: vec![],
-            change: Amount::from_sat(0),
-            fees: PegOutFees {
-                fee_rate: Feerate { sats_per_kvb: 1000 },
-                total_weight: 40000,
-            },
-            destination: destination.clone(),
-            selected_utxos: selected_utxos.clone(),
-            peg_out_amount: Amount::from_sat(10000),
-            rbf: None,
-        };
-
-        dbtx.insert_new_entry(&unsigned_transaction_key, &unsigned_transaction)
-            .await;
-
-        let pending_transaction_key = PendingTransactionKey(Txid::from_slice(&BYTE_32).unwrap());
-
-        let pending_tx = PendingTransaction {
-            tx: transaction,
-            tweak: BYTE_33,
-            change: Amount::from_sat(0),
-            destination,
-            fees: PegOutFees {
-                fee_rate: Feerate { sats_per_kvb: 1000 },
-                total_weight: 40000,
-            },
-            selected_utxos: selected_utxos.clone(),
-            peg_out_amount: Amount::from_sat(10000),
-            rbf: Some(Rbf {
-                fees: PegOutFees {
-                    fee_rate: Feerate { sats_per_kvb: 1000 },
-                    total_weight: 40000,
-                },
-                txid: Txid::from_slice(&BYTE_32).unwrap(),
-            }),
-        };
-        dbtx.insert_new_entry(&pending_transaction_key, &pending_tx)
-            .await;
-
-        let (sk, _) = secp256k1::generate_keypair(&mut OsRng);
-        let secp = secp256k1::Secp256k1::new();
-        let signature = secp.sign_ecdsa(&Message::from_slice(&BYTE_32).unwrap(), &sk);
-        dbtx.insert_new_entry(
-            &PegOutTxSignatureCI(Txid::from_slice(&BYTE_32).unwrap()),
-            &vec![signature],
-        )
-        .await;
-
-        let peg_out_bitcoin_tx = PegOutBitcoinTransaction(OutPoint {
-            txid: TransactionId::from_slice(&BYTE_32).unwrap(),
-            out_idx: 0,
-        });
-
-        dbtx.insert_new_entry(
-            &peg_out_bitcoin_tx,
-            &WalletOutputOutcome::new_v0(Txid::from_slice(&BYTE_32).unwrap()),
-        )
-        .await;
-    }
-
-    #[tokio::test(flavor = "multi_thread")]
-    async fn prepare_db_migration_snapshots() -> anyhow::Result<()> {
-        prepare_db_migration_snapshot(
-            "wallet-v0",
-            |dbtx| {
-                Box::pin(async move {
-                    create_db_with_v0_data(dbtx).await;
-                })
-            },
-            ModuleDecoderRegistry::from_iter([(
-                LEGACY_HARDCODED_INSTANCE_ID_WALLET,
-                WalletCommonInit::KIND,
-                <Wallet as ServerModule>::decoder(),
-            )]),
-        )
-        .await
-    }
-
-    #[tokio::test(flavor = "multi_thread")]
-    async fn test_migrations() -> anyhow::Result<()> {
-        validate_migrations(
-            "wallet",
-            |db| async move {
-                let module = DynServerModuleInit::from(WalletInit);
-                apply_migrations(
-                    &db,
-                    module.module_kind().to_string(),
-                    module.database_version(),
-                    module.get_database_migrations(),
-                )
-                .await
-                .context("Error applying migrations to temp database")?;
-
-                // Verify that all of the data from the wallet namespace can be read. If a
-                // database migration failed or was not properly supplied,
-                // the struct will fail to be read.
-                let mut dbtx = db.begin_transaction().await;
-
-                for prefix in DbKeyPrefix::iter() {
-                    match prefix {
-                        DbKeyPrefix::BlockHash => {
-                            let blocks = dbtx
-                                .find_by_prefix(&BlockHashKeyPrefix)
-                                .await
-                                .collect::<Vec<_>>()
-                                .await;
-                            let num_blocks = blocks.len();
-                            ensure!(
-                                num_blocks > 0,
-                                "validate_migrations was not able to read any BlockHashes"
-                            );
-                        }
-                        DbKeyPrefix::PegOutBitcoinOutPoint => {
-                            let outpoints = dbtx
-                                .find_by_prefix(&PegOutBitcoinTransactionPrefix)
-                                .await
-                                .collect::<Vec<_>>()
-                                .await;
-                            let num_outpoints = outpoints.len();
-                            ensure!(
-                                num_outpoints > 0,
-                                "validate_migrations was not able to read any PegOutBitcoinTransactions"
-                            );
-                        }
-                        DbKeyPrefix::PegOutTxSigCi => {
-                            let sigs = dbtx
-                                .find_by_prefix(&PegOutTxSignatureCIPrefix)
-                                .await
-                                .collect::<Vec<_>>()
-                                .await;
-                            let num_sigs = sigs.len();
-                            ensure!(
-                                num_sigs > 0,
-                                "validate_migrations was not able to read any PegOutTxSigCi"
-                            );
-                        }
-                        DbKeyPrefix::PendingTransaction => {
-                            let pending_txs = dbtx
-                                .find_by_prefix(&PendingTransactionPrefixKey)
-                                .await
-                                .collect::<Vec<_>>()
-                                .await;
-                            let num_txs = pending_txs.len();
-                            ensure!(
-                                num_txs > 0,
-                                "validate_migrations was not able to read any PendingTransactions"
-                            );
-                        }
-                        DbKeyPrefix::PegOutNonce => {
-                            ensure!(dbtx
-                                .get_value(&PegOutNonceKey)
-                                .await
-                                .is_some());
-                        }
-                        DbKeyPrefix::UnsignedTransaction => {
-                            let unsigned_txs = dbtx
-                                .find_by_prefix(&UnsignedTransactionPrefixKey)
-                                .await
-                                .collect::<Vec<_>>()
-                                .await;
-                            let num_txs = unsigned_txs.len();
-                            ensure!(
-                                num_txs > 0,
-                                "validate_migrations was not able to read any UnsignedTransactions"
-                            );
-                        }
-                        DbKeyPrefix::Utxo => {
-                            let utxos = dbtx
-                                .find_by_prefix(&UTXOPrefixKey)
-                                .await
-                                .collect::<Vec<_>>()
-                                .await;
-                            let num_utxos = utxos.len();
-                            ensure!(
-                                num_utxos > 0,
-                                "validate_migrations was not able to read any UTXOs"
-                            );
-                        }
-                        DbKeyPrefix::BlockCountVote => {
-                            let heights = dbtx
-                                .find_by_prefix(&BlockCountVotePrefix)
-                                .await
-                                .collect::<Vec<_>>()
-                                .await;
-                            let num_heights = heights.len();
-                            ensure!(
-                                num_heights > 0,
-                                "validate_migrations was not able to read any block height votes"
-                            );
-                        }
-                        DbKeyPrefix::FeeRateVote => {
-                            let rates = dbtx
-                                .find_by_prefix(&FeeRateVotePrefix)
-                                .await
-                                .collect::<Vec<_>>()
-                                .await;
-                            let num_rates = rates.len();
-                            ensure!(
-                                num_rates > 0,
-                                "validate_migrations was not able to read any fee rate votes"
-                            );
-                        }
-                    }
-                }
-                Ok(())
-            },
-            ModuleDecoderRegistry::from_iter([(
-                LEGACY_HARDCODED_INSTANCE_ID_WALLET,
-                WalletCommonInit::KIND,
-                <Wallet as ServerModule>::decoder(),
-            )]),
-        )
-        .await
-    }
-}
