+++ modules/fedimint-wallet-tests/Cargo.toml
@@ -33,3 +33,7 @@ futures = "0.3.28"
 miniscript = { version = "9.0.2" }
 tokio = { version = "1.26.0", features = ["sync"] }
 tracing = "0.1.37"
+secp256k1 = { version = "0.24.2", features = [ "serde" ] }
+rand = "0.8"
+strum = "0.24"
+strum_macros = "0.24"
